Move define_module to package.py
